requirements.txt
=зшaiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
decorator==5.1.1
discord.py @ git+https://github.com/Rapptz/discord.py@eee65ac39b85edc0353402baa42ca7e72ec286d9
dnspython==2.2.1
frozenlist==1.3.0
glqiwiapi==2.11
idna==3.3
multidict==4.7.6
pydantic==1.9.1
pymongo==4.0.2
pytz==2022.1
requests==2.27.1
typing_extensions==4.1.1
urllib3==1.26.9
validators==0.20.0
websockets==3.4
yarl==1.7.2

